requirements.txt
--pre
ipfabric==5.0.0b2
python-dotenv
deepdiff
